requirements.txt
dash==2.17.1
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
authlib==1.3.0
flask==3.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
gunicorn==21.2.0
